validate_macos_setup.py
#!/usr/bin/env python3
"""
macOS Setup Validation for Sentinel Awakens Trading Interface
Validates the complete setup without opening browser windows.
"""

import sys
import os
import json
import threading
import time
from pathlib import Path

def validate_venv_activation():
    """Validate virtual environment is properly activated"""
    print("🔍 Validating virtual environment...")
    
    # Check if we're in a venv
    in_venv = (hasattr(sys, 'real_prefix') or (getattr(sys, 'base_prefix', sys.prefix) != sys.prefix))
    if not in_venv:
        print("   ❌ Virtual environment not activated")
        print("   💡 Run: source venv/bin/activate")
        return False
    
    # Check venv path
    venv_path = Path(sys.prefix)
    expected_venv = Path.cwd() / "venv"
    if venv_path.resolve() == expected_venv.resolve():
        print(f"   ✅ Using project venv: {venv_path}")
    else:
        print(f"   ⚠️  Using different venv: {venv_path}")
    
    return True

def validate_browser_module():
    """Validate browser module can be imported and configured"""
    print("🌐 Validating browser module...")

    try:
        from browser import init_browser
        import undetected_chromedriver as uc
        from pathlib import Path
        import sys

        print("   ✅ Browser module imported")

        # Test Chrome base directory detection (inline)
        if sys.platform == 'darwin':
            base_dir = Path.home() / 'Library' / 'Application Support' / 'Google' / 'Chrome'
            print(f"   ✅ Chrome base dir (macOS): {base_dir}")

        # Test Chrome options setup
        options = uc.ChromeOptions()
        options.add_argument('--no-first-run')
        options.add_argument('--disable-extensions')
        print("   ✅ Chrome options configured")

        return True

    except Exception as e:
        print(f"   ❌ Browser validation failed: {e}")
        return False

def validate_trading_interface():
    """Validate trading interface can be imported"""
    print("💹 Validating trading interface...")
    
    try:
        from trading_interface import TradingInterface
        print("   ✅ Trading interface imported")
        return True
    except Exception as e:
        print(f"   ❌ Trading interface validation failed: {e}")
        return False

def validate_gui_interface():
    """Validate GUI interface without creating windows"""
    print("🖥️  Validating GUI interface...")
    
    try:
        import tkinter as tk
        from gui_interface import TradingGUI
        
        print("   ✅ GUI interface imported")
        
        # Test tkinter availability
        root = tk.Tk()
        root.withdraw()  # Hide the window
        root.destroy()
        print("   ✅ Tkinter window creation works")
        
        return True
    except Exception as e:
        print(f"   ❌ GUI validation failed: {e}")
        return False

def validate_websocket_functionality():
    """Validate WebSocket functionality"""
    print("🔌 Validating WebSocket functionality...")
    
    try:
        import websocket
        import json
        
        print("   ✅ WebSocket client imported")
        
        # Test WebSocket URL and message format
        ws_url = "wss://matrix.cryptoiq.com/api/sentinel/ws"
        print(f"   ✅ WebSocket URL: {ws_url}")
        
        # Test message serialization
        test_message = {
            "op": "sub",
            "symbols": "BTCUSDT",
            "pv": 100,
            "strings": 10
        }
        serialized = json.dumps(test_message)
        print("   ✅ Message serialization works")
        
        # Test signal format
        test_signal = {
            's': 'BTCUSDT',
            'ts': 1697059200000,
            'pv': 100,
            'cv': 120,
            'delta': 20,
            'strings': 10
        }
        parsed = json.loads(json.dumps(test_signal))
        print("   ✅ Signal format validation works")
        
        return True
    except Exception as e:
        print(f"   ❌ WebSocket validation failed: {e}")
        return False

def validate_config():
    """Validate configuration loading"""
    print("⚙️  Validating configuration...")
    
    try:
        from config import ROLLBIT_URL, SELECTORS, DEBUG_UI_SCAN, DEBUG_NETWORK_SPY
        
        print(f"   ✅ Rollbit URL: {ROLLBIT_URL}")
        print(f"   ✅ Selectors count: {len(SELECTORS)}")
        print(f"   ✅ Debug UI scan: {DEBUG_UI_SCAN}")
        print(f"   ✅ Debug network spy: {DEBUG_NETWORK_SPY}")
        
        return True
    except Exception as e:
        print(f"   ❌ Config validation failed: {e}")
        return False

def validate_main_entry_point():
    """Validate main entry point without running it"""
    print("🚀 Validating main entry point...")
    
    try:
        # Import main module without running it
        import importlib.util
        spec = importlib.util.spec_from_file_location("main", "main.py")
        main_module = importlib.util.module_from_spec(spec)
        
        # Check if main function exists
        spec.loader.exec_module(main_module)
        if hasattr(main_module, 'main'):
            print("   ✅ Main function found")
        else:
            print("   ❌ Main function not found")
            return False
        
        return True
    except Exception as e:
        print(f"   ❌ Main entry point validation failed: {e}")
        return False

def validate_environment_variables():
    """Validate environment variable handling"""
    print("🌍 Validating environment variables...")

    # Test Chrome version override
    original_version = os.environ.get('CHROME_MAJOR_VERSION')
    os.environ['CHROME_MAJOR_VERSION'] = '139'

    try:
        # Just test that the environment variable is set
        version = os.environ.get('CHROME_MAJOR_VERSION')
        if version == '139':
            print("   ✅ Chrome version override works")
        else:
            print("   ❌ Chrome version override failed")
            return False
    except Exception as e:
        print(f"   ❌ Chrome version override failed: {e}")
        return False
    finally:
        if original_version:
            os.environ['CHROME_MAJOR_VERSION'] = original_version
        else:
            os.environ.pop('CHROME_MAJOR_VERSION', None)

    # Test Chrome profile override
    original_profile = os.environ.get('CHROME_PROFILE_DIR')
    os.environ['CHROME_PROFILE_DIR'] = 'Default'

    try:
        # Just test that the environment variable is set
        profile = os.environ.get('CHROME_PROFILE_DIR')
        if profile == 'Default':
            print("   ✅ Chrome profile override works")
        else:
            print("   ❌ Chrome profile override failed")
            return False
    except Exception as e:
        print(f"   ❌ Chrome profile override failed: {e}")
        return False
    finally:
        if original_profile:
            os.environ['CHROME_PROFILE_DIR'] = original_profile
        else:
            os.environ.pop('CHROME_PROFILE_DIR', None)

    return True

def main():
    """Run all validation tests"""
    print("🤖 SENTINEL AWAKENS - macOS Setup Validation")
    print("=" * 60)
    
    validations = [
        validate_venv_activation,
        validate_browser_module,
        validate_trading_interface,
        validate_gui_interface,
        validate_websocket_functionality,
        validate_config,
        validate_main_entry_point,
        validate_environment_variables,
    ]
    
    results = []
    for validation in validations:
        try:
            result = validation()
            results.append(result)
        except Exception as e:
            print(f"   ❌ Validation failed with exception: {e}")
            results.append(False)
        print()
    
    print("=" * 60)
    print("📊 VALIDATION RESULTS:")
    passed = sum(results)
    total = len(results)
    print(f"   Passed: {passed}/{total}")
    
    if passed == total:
        print("   🎉 ALL VALIDATIONS PASSED!")
        print("   ✅ macOS setup is complete and ready")
        print("   🚀 Ready to run: ./run_sentinel.sh")
        print()
        print("   Next steps:")
        print("   1. Run ./run_sentinel.sh")
        print("   2. Wait for browser to open and navigate to Rollbit")
        print("   3. Login manually if needed")
        print("   4. Test manual trades (BUY/SELL buttons)")
        print("   5. Click '🌐 WS Settings' to test WebSocket functionality")
        print("   6. Toggle WebSocket on and test with 'Send Test Signal'")
        return 0
    else:
        print("   ⚠️  Some validations failed. Check issues above.")
        return 1

if __name__ == "__main__":
    sys.exit(main())
