requirements.txt
# Sentinel Awakens Trading Bot - Python Dependencies
# Install with: pip install -r requirements.txt

# Core web automation
selenium>=4.0.0
undetected-chromedriver>=3.5.0

# HTTP requests (if needed for API calls)
requests>=2.28.0

# GUI framework (built-in with Python, but listed for reference)
# tkinter - included with Python standard library

# Additional utilities that might be useful
webdriver-manager>=3.8.0
websocket-client>=1.8.0

# Python 3.13 compatibility (distutils replacement)
setuptools>=65.0.0